refactor: use a default for details in replyAllAction
Simplify the details setup in replyAllAction by using a default details
structure instead of creating two separate clauses. Note that const
only applies to the reference, not the object it points to.

This is both less code and more robust in case we add new options for other
details modifications in the future.

Suggested-by: John Bieling <john.bieling@gmx.de>
Signed-off-by: Jacob Keller <jacob.keller@gmail.com>