refactor: simplify loadOption implementation
In loadOption, we need to get the locally stored option value. This is
implemented by getting the value from the local storage. Then if its
undefined, we look up the value from the DefaultOptions.

According to the documentation, the .get() function can take a dictionary
that specifies what keys to get and what their default values should be if
they don't exist.

Using this, we can simplify the then() block as we guarantee we'll either
the stored value or the default value.

Link: https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/storage/StorageArea/get
Suggested-by: John Bieling <john.bieling@gmx.de>
Signed-off-by: Jacob Keller <jacob.keller@gmail.com>