refactor: remove OptionsList export
The only place where we use OptionsList is in resetAllOptions. The
OptionsList is just a set of keys from DefaultOptions.

Improve code clarity by removing OptionsList and just calling
Object.keys() to get the options in resetAllOptions.

This makes it obvious that the DefaultOptions is the definition of all of
the available options, rather than forcing a reader to check how its
implemented in defaults.js

Suggested-by: John Bieling <john.bieling@gmx.de>
Signed-off-by: Jacob Keller <jacob.keller@gmail.com>